Added: Aut-Update Mod Loader Files
The app on start-up (and when you change the game directory) checks your version of the mod loader against the latest version and notifies you if there's a newer version, prompting you to update. If you choose yes, it attempts to download it and immediately install it. The version in the 'Libs' folder is updated first, then it is copied into the game directory, so that you have the latest files on hand just in-case you need to perform a re-install of the game.